Correct event name
We were attempting to remove a listener for the wrong event.

Correcting this exposes a design mistake: suppose the user pauses
the timeline, then moves the map, then presses play again. In this
case we would like the map to move back to where it was when the
timeline was paused.

However, now we have no way to even know that the map has been
moved while the timeline is paused.

The correct fix is probably to record the position at the point
where the timeline is paused, rather than when the map is moved.