Move Talkie/Leaflet integration to its own file
Also bring in the trick I used on “In Flight” to interrupt map
animations if necessary, which makes the scripted view changes
more reliable.

(For example, if the user moves the map just before a scripted
view change, this will pause the timeline and register an undo
handler. When the timeline is resumed, the undo animation may
still be running when the scripted view change is reached, and
Leaflet will therefore just ignore the scripted view change.
This is not desirable behaviour. Explicitly interrupting the
existing animation like this makes it work as intended.)